Move create repository view outside of components folder
